M2Lib

+++ FireFlameEngine/src/Filters/FLD3DGaussBlurFilter.cpp
@@ -407,7 +407,7 @@ std::vector<float> D3DGaussBlurFilter::CalcGaussWeights()
     }
 
     // Divide by the sum so all the weights add up to 1.0.
-    for (int i = 0; i < weights.size(); ++i)
+    for (size_t i = 0; i < weights.size(); ++i)
     {
         weights[i] /= weightSum;
     }
